fix: cancel button remains active during account creation and further polishing
