+++ packages/mobile/components/StrongholdUnlock.svelte
@@ -3,7 +3,6 @@
     import { unlockStronghold } from '@core/profile'
     import { Button, PasswordInput, Text, HTMLButtonType } from 'shared/components'
 
-    export let subtitle: string | undefined = undefined
     export let busyMessage: string = ''
 
     export let onSuccess: (..._: any[]) => Promise<unknown>
@@ -33,7 +32,7 @@
 </script>
 
 <div class="mb-5">
-    <Text type="p" secondary>{subtitle ?? localize('popups.password.subtitle')}</Text>
+    <Text type="p" secondary>{localize('popups.password.subtitle')}</Text>
 </div>
 <form id="password-popup-form" class="flex justify-center w-full flex-row flex-wrap" on:submit={handleSubmit}>
     <PasswordInput
@@ -45,7 +44,9 @@
         autofocus
     />
     <div class="flex flex-row justify-between w-full space-x-4">
-        <Button outline classes="w-1/2" onClick={handleCancelClick}>{localize('actions.cancel')}</Button>
+        <Button outline classes="w-1/2" onClick={handleCancelClick} disabled={isBusy}>
+            {localize('actions.cancel')}
+        </Button>
         <Button
             classes="w-1/2"
             type={HTMLButtonType.Submit}
