+++ packages/mobile/views/dashboard/drawers/account-switcher/AccountSwitcherDrawer.svelte
@@ -1,16 +1,12 @@
 <script lang="typescript">
-    import { tryCreateAdditionalAccount } from '@core/account'
     import { localize } from '@core/i18n'
-    import { isStrongholdUnlocked } from '@core/profile-manager'
 
-    import { Drawer, StrongholdUnlock } from '../../../../components'
-    import { AccountSwitcherRoute, accountSwitcherRoute, accountSwitcherRouter } from '../../../../lib/routers'
-    import { AccountSwitcher, CreateAccount } from './views'
+    import { Drawer } from '../../../../components'
+    import { AccountSwitcherRoute, accountSwitcherRoute } from '../../../../lib/routers'
+    import AccountSwitcherRouter from './AccountSwitcherRouter.svelte'
 
     export let onClose: () => unknown = () => {}
 
-    let accountAlias: string
-    let color: string
     let title: string
 
     $: $accountSwitcherRoute, setTitle()
@@ -28,38 +24,8 @@
                 break
         }
     }
-
-    async function onCreate(newAccountAlias: string, newColor: string): Promise<void> {
-        accountAlias = newAccountAlias
-        color = newColor
-
-        const isUnlocked = await isStrongholdUnlocked()
-        if (isUnlocked) {
-            await tryCreateAdditionalAccount(newAccountAlias, newColor)
-            onClose()
-        } else {
-            $accountSwitcherRouter.next()
-        }
-    }
-
-    async function onSuccess(): Promise<void> {
-        await tryCreateAdditionalAccount(accountAlias, color)
-        onClose()
-    }
 </script>
 
 <Drawer {onClose} {title}>
-    {#if $accountSwitcherRoute === AccountSwitcherRoute.Switcher}
-        <AccountSwitcher onSwitch={onClose} onAddClicked={() => $accountSwitcherRouter.next()} />
-    {/if}
-    {#if $accountSwitcherRoute === AccountSwitcherRoute.CreateAccount}
-        <CreateAccount {onCreate} onCancel={() => $accountSwitcherRouter.previous()} />
-    {/if}
-    {#if $accountSwitcherRoute === AccountSwitcherRoute.Password}
-        <StrongholdUnlock
-            {onSuccess}
-            onCancel={() => $accountSwitcherRouter.previous()}
-            busyMessage={localize('general.creating')}
-        />
-    {/if}
+    <AccountSwitcherRouter {onClose} />
 </Drawer>
